ai-article-generator/pyproject.toml
[project]
name = "ai-article-generator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "beautifulsoup4>=4.13.5",
    "elevenlabs>=2.16.0",
    "markdown>=3.9",
    "nltk>=3.9.1",
    "openai>=1.108.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
